requirements.txt
# Web UI
streamlit>=1.31.0

# Document Processing
pypdf>=4.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0

# Text Processing & AI
sentence-transformers>=2.2.0

# Output Generation
jinja2>=3.1.0
xlsxwriter>=3.1.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
scikit-learn>=1.3.0
